Extract checkpointer to separate class
